+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.3.2] - ????-??-??
 
+- Track mouse position in the debug box.
 - Fix canvas clean up glitch.
 
 ## [0.3.1] - 2022-04-23